pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastpath-repo"
version = "1.0.0"
description = "Research-grade repository packing system for LLM code analysis with 20-35% improvement in Q&A accuracy"
readme = "README_PYPI.md"
license = {text = "BSD Zero Clause License"}
authors = [
    {name = "FastPath Research Team", email = "contact@fastpath.ai"},
]
maintainers = [
    {name = "FastPath Research Team", email = "contact@fastpath.ai"},
]
keywords = [
    "llm", "code-analysis", "repository-packing", "submodular-optimization", 
    "semantic-search", "ai", "machine-learning", "research", "icse"
]
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tiktoken>=0.5.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "tree-sitter-typescript>=0.20.0", 
    "tree-sitter-go>=0.20.0",
    "tree-sitter-javascript>=0.20.0",
    "tree-sitter-java>=0.20.0",
    "tree-sitter-cpp>=0.20.0",
    
    # System utilities
    "psutil>=5.8.0",
    
    # Text processing
    "markdown>=3.8.2",
    "pygments>=2.19.2",
    
    # CLI and utilities
    "click>=8.0.0",
    "rich>=13.0.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    
    # HTTP requests for repository cloning
    "requests>=2.28.0",
    "gitpython>=3.1.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Utilities",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
research = [
    "scipy>=1.10.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
]
evaluation = [
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.25.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]
all = [
    "fastpath-repo[research,evaluation,dev]"
]

[project.scripts]
fastpath = "packrepo.cli.fastpack:main"
rendergit = "rendergit:main"

[project.urls]
Homepage = "https://github.com/fastpath-ai/fastpath-repo"
Repository = "https://github.com/fastpath-ai/fastpath-repo"  
Documentation = "https://fastpath-repo.readthedocs.io"
"Bug Reports" = "https://github.com/fastpath-ai/fastpath-repo/issues"
"Research Paper" = "https://arxiv.org/abs/2024.fastpath"
"ICSE 2025" = "https://conf.researchr.org/track/icse-2025/icse-2025-research-track"
"PyPI Package" = "https://pypi.org/project/fastpath-repo/"

[tool.setuptools.packages.find]
include = ["packrepo*"]
exclude = ["tests*", "docs*", "examples*", "scripts*"]

[tool.setuptools]
py-modules = ["rendergit"]

[tool.setuptools.package-data]
"packrepo" = ["configs/*.yaml", "configs/*.json"]
"packrepo.evaluator" = ["prompts/*.md", "configs/*.json"]

# Development tools configuration
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["packrepo"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_reexport = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests", "packrepo/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-config",
    "--strict-markers",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "research: marks tests for research validation",
]
